infrastructure/databricks-apps/nyctaxi-app/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.65.0",
    "fastapi>=0.117.1",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.11.9",
    "uvicorn>=0.36.0",
]

